Update postgres.md
Eliminating marketing language for CrunchyDB as Credativ our partner in our blog w/ them suggests another operator and, in any case, the operator need not be featured - instead the reader likely wants to know a little bit about why and mostly about how they will jointly deploy.  QUESTION - we refer to the OpenEBS.ci reference for running postgreSQL however recently OpenEBS.ci has been reconfigured.  ALSO we say here that cStor volume resize is not seamless - perhaps we should link here to how it is done even in a non seamless way?  Or maybe we considered that and decided not to do so?