Update ugndm.md
Updated note to mention that ndm daemon set should be restarted once the changes in performed on ndm config where OpenEBS is already installed and running. This change is added for both include and exclude filter. 
Also added a note that config map changes are global and regex support is not present.

Signed-off-by: Ranjith R <ranjith.raveendran@mayadata.io>