fix(localpv): note to avoid nodeName in podspec
Fixes https://github.com/openebs/openebs/issues/2915

Using `nodeName` in Pod spec results in scheduler skipping
the step of assigning a node to Pod and inturn not setting
the node details on PVC. This will result in PVC being stuck
in pending state, when `waitForFirstconsumer` is specified
in the corresponding Storage Class.

This commit adds a note to docs to avoid specifying nodeName
in pod spec.

Signed-off-by: kmova <kiran.mova@mayadata.io>