feat(alpha): refactor alpha features page
Refactor the alpha features page to redirect to
corresponding repository documentation. This will
avoid getting the docs out of sync from features
that are under active development.

Also included the additional alpha features that
are currently active.

Removed the jiva from alpha section as the development
has not happened on it for couple of releases now.
Will add it back when it starts to gain traction.

Signed-off-by: kmova <kiran.mova@mayadata.io>